Begin work on direct login
This will remove the need to have the router page open as well as the extensoin's page.

Methods to get verification tokens and hashing have been added. A new saveAjaxDataDirect has also been added.

More work is needed to include the session ID in XHR requests. This is stored as a cookie.