Clean up router module
The router module should be separated into another npm package. This commit begins work on doing so.

In order to separate the router api into another module, the dependency on chrome's api must be removed. For instance, the storage api is being replaced with a more general config class.

Any utility functions that are needed are being added to the router module.

Additionally, all tab-related functions are being removed.