Get all messages at once instead of one page at a time
Features such as labelling cannot be implemented when pages are gotten one at a time because the router API does not allow getting a single message by ID. Not being able to retreive a message by ID means we can't display a list of messages that belong to a particular label without checking every page of every box type. Thus, even though getting messages for only the current page is faster, we must get all the messages at once.